split jobs router into multiple jobs router modules; simplify status routes for jobs router
