+++ diracx-routers/pyproject.toml
@@ -48,14 +48,14 @@ types = [
 ]
 
 [project.entry-points."diracx.services"]
-jobs = "diracx.routers.job_manager:router"
+jobs = "diracx.routers.jobs:router"
 config = "diracx.routers.configuration:router"
 auth = "diracx.routers.auth:router"
 ".well-known" = "diracx.routers.auth.well_known:router"
 
 [project.entry-points."diracx.access_policies"]
-WMSAccessPolicy = "diracx.routers.job_manager.access_policies:WMSAccessPolicy"
-SandboxAccessPolicy = "diracx.routers.job_manager.access_policies:SandboxAccessPolicy"
+WMSAccessPolicy = "diracx.routers.jobs.access_policies:WMSAccessPolicy"
+SandboxAccessPolicy = "diracx.routers.jobs.access_policies:SandboxAccessPolicy"
 
 # Minimum version of the client supported
 [project.entry-points."diracx.min_client_version"]
